refactor: перенеси логику добавления/удаления друзей в UserStorage
