+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -15,4 +15,8 @@ public interface UserStorage {
     User getUserById(int id);
 
     boolean containsUserById(int id);
+
+    void addFriend(int userId, int friendId);
+
+    void deleteFriend(int userId, int friendId);
 }
